Move User retrieval logic to UserStore
