+++ db/user_store.go
@@ -2,7 +2,8 @@ package db
 
 import (
 	"database/sql"
-	"log"
+	"errors"
+	"fmt"
 )
 
 type UserStore struct {
@@ -34,16 +35,39 @@ func (s *UserStore) SaveUserScore(user *User) error {
 	return err
 }
 
-func (s *UserStore) GetUserCount() int {
+func (s *UserStore) GetUserCount() (int, error) {
 	var count int
 	res := s.DB.QueryRow("SELECT COUNT(*) FROM users")
 
 	err := res.Scan(&count)
 
-	if err != nil {
-		log.Fatal(err)
-		return -1
+	return count, err
+}
+
+func (s *UserStore) GetUserByID(userID string) (*User, error) {
+	user := &User{}
+	res := s.DB.QueryRow("SELECT user_id, display_name, sync_code, pets FROM users WHERE user_id = ?", userID)
+
+	if err := res.Scan(&user.UserID, &user.DisplayName, &user.SyncCode, &user.PetCount); err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return nil, fmt.Errorf("no user with id: %v", userID)
+		}
+		return nil, err
 	}
 
-	return count
+	return user, nil
+}
+
+func (s *UserStore) GetTotalPetCount() (int, error) {
+	var count int
+	res := s.DB.QueryRow("SELECT SUM(pets) FROM users")
+
+	err := res.Scan(&count)
+
+	if err != nil {
+		return 0, err
+	}
+
+	return count, nil
+
 }
